fix: replace disabled button with info border
