+++ lib/common/common.dart
@@ -1,7 +1,7 @@
 export 'circle_indicators.dart';
-export 'disabled_button.dart';
 export 'footer_links.dart';
 export 'format_text.dart';
 export 'global_header.dart';
+export 'info_border.dart';
 export 'primary_button.dart';
 export 'styles.dart';
